requirements.txt
Django==4.0.5
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
